refactor: SideBar 리팩터링 및 주석 추가
