+++ components/common/Sidebar/styles.tsx
@@ -3,24 +3,27 @@ import styled from 'styled-components';
 export const Container = styled.section<{ width: number }>`
   position: relative;
   height: 100vh;
-  width: ${({ width }) => width / 10}rem;
   background-color: #f7f6f3;
 
   flex-grow: 0;
   flex-shrink: 0;
+  overflow: hidden;
 
   pointer-events: none;
   user-select: none;
 
-  min-width: 19rem;
-  max-width: 48rem;
+  width: ${({ width }) => width / 10}rem;
+  // ! 최대 너비를 제한하는 코드
+  /* max-width: 48rem; */
+  /* min-width: 19rem; */
 
   padding-top: 1.5rem;
   padding-left: 1.5rem;
+
+  // ! css 추후 사용
   /* display: flex; */
   /* flex-direction: row; */
   /* font-size: 5rem; */
-  /* max-width: 30rem; */
   /* padding: 1.5rem; */
 `;
 
@@ -28,12 +31,13 @@ export const SideBarContents = styled.div`
   display: flex;
 
   pointer-events: auto;
+
+  flex-grow: 0;
+  flex-shrink: 0;
+
+  // ! css 추후 사용
   /* flex-direction: row; */
-
   /* height: 100vh; */
-
-  /* flex-grow: 0; */
-  /* flex-shrink: 0; */
   /* flex-direction: column; */
 `;
 
@@ -41,7 +45,7 @@ export const SideBarMenuContainer = styled.div`
   flex: 1;
 `;
 
-export const SideBarResizer = styled.div`
+export const SideBarResizer = styled.div<{ isVisible: boolean }>`
   position: absolute;
   right: 0;
   top: 0;
@@ -52,10 +56,12 @@ export const SideBarResizer = styled.div`
   cursor: col-resize;
 
   &:hover {
-    width: 4px;
     width: 3px;
-    background: #c1c3c5b4;
+    background-color: #c1c3c5b4;
   }
+
+  // ! css 추후 사용
+  /* background-color: ${({ isVisible }) => (isVisible ? '#c1c3c5b4' : 'transparent')}; */
   /* background-color: red; */
 
   /* flex-basis: 1.2rem; */
@@ -65,7 +71,6 @@ export const SideBarResizer = styled.div`
   /* width: 100px; */
   /* width: 100px; */
   /* width: 5rem;
-  background-color: red;
   border-right: red 1px solid;
   height: 100%; */
   /*
